pyproject.toml
[project]
name = "custom-skill-agent"
version = "0.1.0"
description = "Framework-agnostic skill system for AI agents with progressive disclosure"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai>=0.0.30",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "openai>=1.0.0",
    "sqlalchemy[asyncio]~=2.0.36",
    "asyncpg~=0.30.0",
    "alembic~=1.14.0",
    "pgvector~=0.3.6",
    "redis[hiredis]~=5.2.0",
    "fastapi~=0.115.0",
    "uvicorn[standard]~=0.32.0",
    "python-jose[cryptography]~=3.3.0",
    "bcrypt~=4.2.0",
    "python-multipart~=0.0.17",
    "langfuse>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true

[dependency-groups]
dev = [
    "fakeredis[lua]~=2.26.0",
    "httpx>=0.27.0",
    "mypy>=1.19.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.8",
]
